Move OLE sources to Core assembly (#12859)
Move OLE sources to Core assembly

This moves and renames files to build in the Core assembly.

- Adds a few more scenarios missed to IOleServices
- Fix nits from last PR
- Adds global suppressions for trimming for moved code
- Rename a few other files that were moved previously

It is sort of odd that we force GetImage through HBITMAP instead of using the serialized System.Drawing.Bitmap. This tweaks the colors (at least transparent ones) as Clipboard_SetImage_InvokeBitmap_GetReturnsExpected shows. Not sure this is the right behavior, but it is what we have currently.