Add matching options to TypeNameComparer
Added matching options (`TypeNameComparison`) for `TypeNames` to better align with what we support in our Nrbf code.

Splits BinaryFormatUtilities and moves the Core only parts to the Core test assembly. Had to move other code and make tweaks to the test project to make this possible.

Starts some tests for CoreNrbfSerializer.