Ensure `ComManagedStream` seekable wrapper stream has position of 0 (#12953)
Until .NET 5, `ComManagedStream` (then `GPStream`) would wrap a
non-seekable stream in a `MemoryStream` using the `MemoryStream(Byte[])`
constructor. This results in a stream with a position of 0.

dotnet/runtime commit 136527537e6 (Improve perfromance for loading from
Stream on Windows (dotnet/corefx#31142), 2018-07-20) updated this logic
to instead use `CopyTo` to populate the wrapping `MemoryStream`. This
results in a stream with a non-zero position.

It seems that this non-zero position causes some issues in downstream
code when using `Image.FromStream` to load `.emf` and `.wmf` files,
resulting in `LoadGdipImageFromStream` returning a status of 2 and thus
an exception.

Seek the wrapping `MemoryStream` back to the beginning after copying the
source stream into it to prevent this exception.

Fixes #12951