Start moving tests to single test subfolder
This starts moving the tests to a single test root folder. That plays better with `Directory.*` files, `.editorconfig` files, etc. (where we can better align these with WPF). This will also help a bit with `MAX_PATH`.

Moved tests are picking up the default Microsoft Shared key. The intent, however, is to move all tests to the Open Shared key when the move is complete.