Property grid entry for ShortCut keys should not be up/down scrollable (#13280)
ToolStripMenuItem.ShortcutKeys property type (Keys enum) has a TypeConverter that reports that it has a set of standard values. This is not entirely accurate. Keys type has a set of standard values, but the ShortcutKeys property does not, Shortcuts are combinations or keys and modifiers that follow certain rules, it's not practical to present all combinations for user to step through with up/down keys or with mouse wheel. We need to disable scrolling through potential values in the property grid.
This behavior regressed when we added Keys.None value to the list of standard values - #8401. None wasn't in the list of standard values on NETFX. This was done to localize "None" (as "(none)") in the property grid. We should keep this value in the type converted tables for localization purposes. But we should remove it from the list of standard values. When the current value of the shortcut is not one of the standard values, property grid entry does not enter the "scroll through the standard items" functionality, which we want to block again.