Add memory pressure management to AgileComPointer
Enhance the `AgileComPointer<TInterface>` class by introducing a `_memoryPressure` field and updating the constructor to accept a `memoryPressure` parameter. This allows for better memory management by informing the garbage collector about native memory usage.

In the `HtmlElement` class, define a `MemoryPressure` constant and update its constructor to pass this value to `AgileComPointer`, improving garbage collection efficiency when creating multiple `HtmlElement` instances.

Tried creating a cached version of the `AgileComPointer`, but that didn't help in this case as items in the linked repro were always new instances.

Fixes: #13195