Add _isReleasingDataSource to prevent unnecessary operations on CurrentCell when changing or releasing DataSource (#13320)
Fixes #13304

Root cause
Regression introduced in PR #4637

When closing the dialog in edit mode, DataSource is set to null, and then CurrentCell = null is set. Then unnecessary method EndEdit of the SetCurrentCellAddressCore is called.

Proposed changes
Add _isReleasingDataSource in DataGridView and set it to null in property DataSource before setting CurrentCell = null to prevent the EndEdit method of the SetCurrentCellAddressCore from being called
Customer Impact
When the DataGridView is in editing mode, its dialog box can be closed normally
Regression?
Yes
Risk
Minimal