Throw ExternalException instead of OutOfMemory
GDI+ isn't particularly good at returning errors when it is unable to create internal objects. There are many cases where object creation will fail due to invalid input and higher level code will get a null and turn it into `Status.OutOfMemory`.

As this is a frequent cause of confusion, we're changing this to `ExternalException`, which is already thrown in other code paths. We had already changed the exception message.

Fixes #12627