Move TestAccessors to extension properties
Cleaner syntax. Also update exception handling to be more useful.

This is to better align with what I'm putting in the SDK repo. Hopefully I can eventually move this upstream to some of our shared test infra.

It does make it a little bit more awkward for VB, but not unusable.